fix pointer type

+++ source/dnode/vnode/inc/vnode.h
@@ -110,7 +110,7 @@ int     tqReadHandleAddTbUidList(STqReadHandle *pHandle, const SArray *tbUidList
 int32_t tqReadHandleSetMsg(STqReadHandle *pHandle, SSubmitReq *pMsg, int64_t ver);
 bool    tqNextDataBlock(STqReadHandle *pHandle);
 int32_t tqRetrieveDataBlock(SArray **ppCols, STqReadHandle *pHandle, uint64_t *pGroupId, int32_t *pNumOfRows,
-                            int32_t *pNumOfCols);
+                            int16_t *pNumOfCols);
 
 // need to reposition
